review config and performance modules

+++ common_functions.c
@@ -94,9 +94,9 @@ long int get_nanoelapsed_llu(long int t1)
 	agios_gettime(&t2);
 	return (get_timespec2llu(t2) - t1);
 }
-long double get_ns2s(long int t1)
+double get_ns2s(long int t1)
 {
-	long double ret = t1 / 1000.0;
+	double ret = ((double)t1) / 1000.0;
 	ret = ret / 1000.0;
 	return ret / 1000.0;
 }
